relocate files into database
