+++ .idea/modules.xml
@@ -3,6 +3,8 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Swoop.iml" filepath="$PROJECT_DIR$/Swoop.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Swoop.iml" filepath="$PROJECT_DIR$/Swoop.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
